# Conflicts:
#   examples/bluetooth/bluedroid/ble/ble_ancs/sdkconfig.defaults.esp32c2
#   examples/bluetooth/bluedroid/ble/ble_compatibility_test/sdkconfig.defaults.esp32c2
#   examples/bluetooth/bluedroid/ble/ble_eddystone/sdkconfig.defaults.esp32c2
#   examples/bluetooth/bluedroid/ble/ble_hid_device_demo/sdkconfig.defaults.esp32c2
#   examples/bluetooth/bluedroid/ble/ble_ibeacon/sdkconfig.defaults.esp32c2
#   examples/bluetooth/bluedroid/ble/ble_spp_client/sdkconfig.defaults.esp32c2
#   examples/bluetooth/bluedroid/ble/ble_spp_server/sdkconfig.defaults.esp32c2
#   examples/bluetooth/bluedroid/ble/ble_throughput/throughput_client/sdkconfig.defaults.esp32c2
#   examples/bluetooth/bluedroid/ble/ble_throughput/throughput_server/sdkconfig.defaults.esp32c2
#   examples/bluetooth/bluedroid/ble/gatt_client/sdkconfig.defaults.esp32c2
#   examples/bluetooth/bluedroid/ble/gatt_security_client/sdkconfig.defaults.esp32c2
#   examples/bluetooth/bluedroid/ble/gatt_security_server/sdkconfig.defaults.esp32c2
#   examples/bluetooth/bluedroid/ble/gatt_server/sdkconfig.defaults.esp32c2
#   examples/bluetooth/bluedroid/ble/gatt_server_service_table/sdkconfig.defaults.esp32c2
#   examples/bluetooth/bluedroid/ble/gattc_multi_connect/sdkconfig.defaults.esp32c2
#   examples/bluetooth/bluedroid/ble_50/ble50_security_client/sdkconfig.defaults.esp32c2
#   examples/bluetooth/bluedroid/ble_50/ble50_security_server/sdkconfig.defaults.esp32c2
#   examples/bluetooth/bluedroid/ble_50/multi-adv/sdkconfig.defaults.esp32c2
#   examples/bluetooth/bluedroid/ble_50/peroidic_adv/sdkconfig.defaults.esp32c2
#   examples/bluetooth/bluedroid/ble_50/peroidic_sync/sdkconfig.defaults.esp32c2

+++ components/bt/host/bluedroid/btc/profile/std/include/btc_gatts.h
@@ -55,7 +55,7 @@ typedef union {
     struct create_attr_tab_args{
         esp_gatt_if_t gatts_if;
         uint8_t srvc_inst_id;
-        uint8_t max_nb_attr;
+        uint16_t max_nb_attr;
         esp_gatts_attr_db_t *gatts_attr_db;
     }create_attr_tab;
 
@@ -149,8 +149,8 @@ typedef struct {
     esp_bt_uuid_t svc_uuid;
     bool        is_tab_creat_svc;
     bool      is_use_svc;
-    uint8_t   num_handle;
-    uint8_t   handle_idx;
+    uint16_t   num_handle;
+    uint16_t   handle_idx;
     uint16_t handles[ESP_GATT_ATTR_HANDLE_MAX];
 } esp_btc_creat_tab_t;
